top 5 best selling products feature

+++ src/main/java/com/module5/team2/dto/response/OrderItemResponse.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 @Data
 @Builder
 public class OrderItemResponse {
-    private Long productId;
+    private int productId;
     private String productName;
     private Integer quantity;
     private BigDecimal unitPrice;
